refactoring structure of files
