+++ tests/e2e/default/contacts/new-lineage-structure.wdio-spec.js
@@ -3,7 +3,7 @@ const chai = require('chai');
 const contactPage = require('../../../page-objects/contacts/contacts.wdio.page');
 const loginPage = require('../../../page-objects/login/login.wdio.page');
 const commonPage = require('../../../page-objects/common/common.wdio.page');
-const sentinelUtils = require('../../sentinel/utils');
+const sentinelUtils = require('../../../utils/sentinel');
 
 const centerName = 'Franklin';
 const centerContact = 'Center Contact';
