+++ tests/e2e/default/db/db-sync.wdio-spec.js
@@ -1,6 +1,6 @@
 const commonElements = require('../../../page-objects/common/common.wdio.page');
 const utils = require('../../../utils');
-const sentinelUtils = require('../../sentinel/utils');
+const sentinelUtils = require('../../../utils/sentinel');
 const loginPage = require('../../../page-objects/login/login.wdio.page');
 const reportsPage = require('../../../page-objects/reports/reports.wdio.page');
 const chai = require('chai');
