+++ tests/e2e/protractor/sms/registration-by-sms.js
@@ -1,5 +1,5 @@
 const utils = require('../../../utils');
-const sUtils = require('../../sentinel/utils');
+const sUtils = require('../../../utils/sentinel');
 const commonElements = require('../../../page-objects/common/common.po.js');
 const helper = require('../../../helper');
 const moment = require('moment');
